Changed the block/callback requirement. Originally I had planned to require that nonblocking api calls supply a callback for any results returned by the buffer consumers.  I have removed this requirement since it is safe to assume that consumers have sole responsibility for handling those objects and no return is required.
If a callback is supplied then it will work the same when block=False.

If a callback is not supplied for async API call a warning is merely
logged via the logger.

Finally, I have split queued items (consumer handled) from unqueued
items in the buffer.  This means that the buffer can return after
parsing is completed with results that aren't handled by the consumer.

This list of queued items can be over-ridden by the smpapi calling
thread.